Fixes a problem reported by Chris Hoffman that was breaking his PPC nightly regression tests.
unsetEnvironment() was not turning off
all of the variables that the config file might have set.  This, in
turn, broke cross-builds.

When I fixed cross builds, I found that the value of CLASSPATH_ROOT
was now broken on the host machine!

We now have explicit variables, HOST_CLASSPATH_ROOT and
TARGET_CLASSPATH_ROOT.    We use both during our build.