Second part of fix to #1511447.   In a heavily mutating, but non-allocating context, it is possible that generational collectors will exhaust the available virtual memory for metadata through the generation of a large number of remembered set entries.  Since there is no allocation, there are no checks for GC.  GC cannot be triggered from within a write barrier, so there's a problem.
Fortunately this problem was solved years ago for RefCount et al by introducing the idea of asyncrhonous GC, which allows the barrier initiate an "asynchronous" GC when space is low.  The GC is performed at the next GC safe point (not within the barrier).  The change required to make this work for the generational collectors was trivial.