More cleanup of PPC jni code to replace preprocessor directives with static final booleans.  Tested on AIX32/AIX64 only.
The more I revisit this code, the more convinced I am that we need
to refactor to have more virtual methods, and then split all of the
code dealing with the gory details of the calling
conventions into 3 ABI-specific subclasses of an abstract PPC superclass.
As it is, this code is horrible.