Eliminate RVM_WITH_OPT_COMPILER and RVM_WITH_ADAPITVE_SYSTEM from src, src-ia32, and src-ppc trees.  This was enabled by changing the build system so that we generate & java compile the complete (modulo architecture) source tree for all configurations, but then filter the set of classes put in jksvm.jar to weed out opt, osr and adaptive packages when they are not actually included in the build.
Build times become slower for prototype images (since we
generate/compile the opt compiler), but the actual size and contents
of the final built image is unchanged.

There were a few places where to get this to all hang together I took
some shortcuts and declared fields of VM_RuntimeCompiler and
VM_Thread to be of type Object instead of their proper types.  This
was necessary to get the BootImageWriter to work when the proper type
of the field was not included in the smaller (prototype) bootimage.
In the case of VM_RuntimeCompiler, this can be fixed by restructuring
as an abstract class with compiler-specific subclasses (like
VM_BootImageCompiler). For VM_Thread, the 'right' way to structure the
code is less obvious.

Tested with prototype, prototype-opt, and development images on
Linux/IA32, PowerPC/AIX32 and PowerPC/AIX64. Will do more complete
testing overnight and fix any lingering problems as soon as possible
in the morning.