AnnotatedElement.isAnnotationPresent has semantics that require instantiation of annotation objects to correctly obey inheritance.
The bootimage tool needs to determine whether an annotation is present as part of compilation of methods and this occurs before the system is ready to instantiate objects. So rather than modifying the semantics of isAnnotationPresent when not running under a VM provide an alternative path to get access to declared annotations (NOTE: inheritance not supported). Re-route all usages of isAnnotationPresent through this mechanism as none of the annotations are @Inherited anyways. Homogenize the method used to determine if an annotation is present to has<AnnotationName>Annotation() and pull up to VM_AnnotatedElement where sensible.