Remove xalan test.
It was never integrated into the ant testing system as dacapo xalan should test most of the same things. The exception being classloading when a bytearray is supplied but no name. The new test system has a test for this in test.org.jikesrvm.basic.core.classloading.TestLoadingWithoutName - which makes this test redundant. (If this is not the case then we need to revive xalan and integrate it into the new system)