Enable constant folding for Tib values.
To do this there needed to be two bugs fixed within getObjTib.

The first path handled null constants that should not be present (as it is dead code) BUT it only did this when assertins were enabled which means that when assertions were disabled the null constant would fall through to else block and cause a failure.

The second bug was the assumption that constant values have a resolved type which did not seem to hold so if it is guarded with an isResolved it seemed to cause no more problems.

Fix for: RVM-16