RVM-476 : Make Lock allocation less dependent on Processors/Green Threading.
Patch from Filip updated to match svn head.  This removes a fairly complex (but not actually enabled) mechanism for re-balancing free locks between the Processors and instead uses a simple one element freeLock cache on each RVMThread to reduce contention on the global pool of free Locks.  Fil's initial experiments showed that the one element cache was sufficient to avoid serious contention on the global lock free list.  More complex schemes could be devised, but until we actually see a problem we'll stay simple.  I'll monitor performance runs for the next day or so to see if this is an issue...

This change should also remove the intermittent jbb2000/2005 crash (1-3 times a week in the nightly runs) that has been happening because without re-balancing (which was disabled), the system was unable to effectively recycle freed locks when allocation/free operations were not evenly distributed among the Processors.