Bump up threshold from 5 seconds to 10 seconds before warning that a virtual processor may be stuck.
This may help ppc-aix regression runs on piccolo avoid load-induced failures.  If we weren't about to eliminate green threading, I would instead want to investigate why there has been a large increase in these kinds of warnings on AIX recently.  The crontab/load on the machine hasn't been changed in months, so a spike in these warnings does indicate some kind of problem in the system.