RVM-1036 : Treat checkstore checks similarly to array bound checks and null checks when creating the initial IR from bytecode.
The old treatment of checkstore checks was incorrect: checkstore checks were never inserted in inlined code because the respective variable was not set in createChildContext(..).  It would have been possible to fix the bug in createChildContext(..) but I don't see a reason to treat checkstore checks differently than the other kinds of checks.