Change implementation of soft handshakes so that CollectorThreads never take part in soft handshakes.
The previous implementation of soft handshakes deadlocked as soon as a CollectorThread was involved in the soft handshake. This happened because CollectorThreads do not yield and therefore cannot acknowledge soft handshakes.

Making soft handshakes work for CollectorThreads might require a larger rework of the soft handshake implementation. There is currently no need to do this because all implementations of SoftHandshakeVisitor in the Jikes RVM were excluding CollectorThreads anyway.